Move question/answer form-rendering tests to a focused file (Codex review)
The form-field rendering cases are a discrete block hardening questionTextForm/
answerTextForm, so extract them from the 743-line admin-questions template
suite into test/templates/admin/question-forms.test.ts. Per AGENTS.md's
"400-line limit applies to test files", a narrow file means future questions.ts
mutation runs execute only these cases per mutant, not the whole template suite.
